scripts/requirements.txt
requests
ollama
pyyaml
streamlit
chromadb
packaging
watchdog

